gobi3k-sdk: Use lock instead of trylock
Trylock appears to be used here as part of some kind of fancy-but-unnecessary
deadlock-detection system; we're better off just writing deadlock-free code to
begin with.

The trylock-using code was losing events; if it tried to get the lock and
failed, it'd drop the incoming event on the floor, which means response latency
would nondeterministically spike up to the max timeout (thus the test failures
with timeout = 1000 instead of 100). If we just wait until the lock is free,
we'll never discard events. Go us.

(This also makes the timeout unnecessary in most cases, but I have not removed
it in this CL yet.)

BUG=chromium-os:13475
TEST=network_3GStressEnable

Signed-off-by: Elly Jones <ellyjones@google.com>
Signed-off-by: Jason Glasgow <jglasgow@google.com>

Review URL: https://chromereviews.googleplex.com/2727013