qmi: fix up type handling for fields
The fields 'type' isn't	the actual C member type, it's either
the C member type (QC calls this a "STD" type) or it's an enum
type.  If it's an enum type, use that type in the struct.  Also
handle fixed-length ANSI string types, which are essentially a
pointer to a string with the length encoded the previous struct
member.