dms: new opaque type to return the results of the GET_IDS request
We do not want to explicitly list every possible return value in the
`qmi_client_dms_get_ids_finish()', as that would break compatibility if we ever
need to add a new value being returned. Instead, we will create opaque types for
the return of each operation, with specific getters for the values returned.