ctl: new opaque type to return the results of the VERSION_INFO request
We do not want to explicitly list every possible return value in the
`qmi_client_ctl_get_version_info_finish()' method, as that would break
compatibility if we ever need to add a new value being returned. Instead, we
will create opaque types for the return of each operation, with specific getters
for the values returned.