cli,wds: new `--wds-follow-network' switch
When `--wds-start-network' is run along with `--wds-follow-network', the
connection will be opened and kept tracked as long as it is alive. Once the
modem gets disconnected, the `qmicli' process will end.

When `--wds-start-network' is NOT run with `--wds-follow-network', the program
will exit immediately. In this case, the user should keep track of the printed
packet data handle and should also not release the CID; so that she can then
issue a `--wds-stop-network' command with the proper CID and packet data handle.