device: when opening, run several version checks if requested
When early requests arrive to the modem, when it is still internally
initializing, the modem may completely discard the requests and our
actions get timed out.

In order to try to avoid that, instead of a single version-check with
N seconds timeout, we'll try N version-check requests with 1s timeout.