qmi-codegen: use `g_array_set_clear_func()' to clear array contents
Bumped required glib version to 2.32. Actually... there is no real issue with
keeping 2.28 as minimum required version, as for now there's no array element
which requires clearing, but anyway, given that QMI is to be used in quite
recent kernels, it shouldn't be a big issue to require a recent glib as well.