qmi-message: fix minimum size of buffer needed to read a QMI message
The buffer must contain *at least* the initial 1-byte marker plus the length
reported by the QMUX header. The minimum size check was wrong for 2 bytes, which
could cause errors when trying to decode a message without all bytes.

Can easily be triggered using 1 for the BUFFER_SIZE in QmiDevice.