libqmi-glib,qmi-codegen: timed out operations will issue an ABORT message
Messages can now be tagged with a special 'abort' keyword, so that whenever the
message times out we issue a new ABORT command to cancel the specific timed out
request.

This support is currently only available for the NAS and WDS services, which are
the ones supporting ABORT for their long-running operations.