mbim-codegen,libmbim-glib: improve general array and variable reading
In the general variable reading part, we no longer need specific methods for
'command' or 'command-done' message types.

Also, when reading arrays, we will specify separately the variable with the
array size and the variable with the actual array contents. Main reason for this
is that it seems they don't have to be one after the other.

The arrays of guint32 will be '0' terminated. This is useful for parameters
where '0' is not an expected value. In general, the size of arrays of guint32
values is given in the output 'size' variable when the array is read.