qmi-network: Fix grabbing output from lines with spaces
The qmi-network script used awk with a field separator of ":". The
output looked like key: 'value'. The second field in this case includes
a space at the beginning. This was making the 'status' command fail.

Now, we account for the space. It is now also not dependent on awk
(this can matter on embedded systems).