libqmi-glib,utils: fix reading sized uint from Big Endian buffers
When reading a sized Little Endian uint from a buffer, we just need to read
N bytes from the buffer and place them at the beginning of a new 8-byte buffer,
(as if it were a guint64) and finally convert from LE to HE.

But when reading a sized Big Endian uint from a buffer, the N bytes that we read
from the beginning of the buffer need to be placed at the *end* of the new
8-byte buffer before converting from BE to HE.

Sized integers are used to read e.g. 6-byte integers, like part of the 'Device
Time' TLV in the 'DMS Get Time' call.

https://bugs.freedesktop.org/show_bug.cgi?id=65078