mbim-codegen: handle optional fields flagged with 'available-if'
A given MBIM message may contain fields which can be processed in different ways
depending on the value of another field. This is the case for e.g. "SMS Read"
Query operations, where the list of messages may be formatted as either PDU or
CDMA, depending on the value of the 'Format' field.

So, we now support to have fields configured like this:

   { "name"             : "Something",
     "format"           : "string" ,
     "available-if"     : { "field"     : "ConditionField",
                            "operation" : "==",
                            "value"     : "0" } },
   { "name"             : "OtherSomething",
     "format"           : "string" ,
     "available-if"     : { "field"     : "ConditionField",
                            "operation" : "==",
                            "value"     : "1" } },

Meaning that:
 * Field "Something" will be given if "ConditionField" == 0
 * Otherwise, Field "OtherSomething" will be given if "ConditionField" == 1

Note that in the message parser, BOTH fields will be provided, but only one
will contain valid data.