libqmi-glib,qmi-proxy: only allow clients running as root
Check the uid of the remote user; and only allow connections if the user is
root. Also, make sure that only the root user can actually create the qmi-proxy
daemon.

Looking at the user id to see whether it will have privileges to open the
/dev/cdc-wdm port port is probably not the best option; but the cleanest and
easiest way for now, given that the qmi-proxy is launched on-demand and not
always running.