uim: fix 'File ID' TLV format
The 'File ID' TLV expects the File path given as a sequence of BE-encoded 16-bit
numbers; but the actual array is given as a 8-bit encoded array (i.e. size of
the array should be in bytes).

So instead of a 'File ID' like this:
  46:6F:02:00:3F:20:7F

it should be like this:
  46:6F:04:00:3F:20:7F