qmi-network: use dot instead of 'source' to load the profiles
The qmi-network script doesn't work properly because it uses a bash-specific
"source" command. In Ubuntu, /bin/sh is dash. The solution is to just change
"source" to a dot, which is the proper way to do a "source" in POSIX sh.

Bug report and original fix by Heath Kehoe <heath@digitalartefacts.com>