mbimcli: check status code in command-done message before parsing response
This patch adds missing mbim_message_command_done_get_result() checks in
mbimcli to ensure that the status code in a command-done message reports
no error before parsing further parsing the response. It also adds
missing mbim_message_unref() calls to ensure that the response is freed
in case of an error.