libmbim-glib,proxy: don't re-configure global indications if merged unchanged
Currently we're defaulting to enable the standard services list; so whenever
clients want to re-configure their standard services list (e.g. just enabling
some of the indications), we just end up having the full standard list again.

So, avoid multiple re-configures in the device, just compare and only send the
new setup to the device if the global list changes.