libmbim-glib,proxy: add a configure flag to set the user ID of MBIM proxy
Currently, the MBIM proxy process assumes that it is run as root user and
that all incoming client connection users are also root.
However, it's not always preferable to run the MBIM proxy as root for
security reasons. On some platforms, the MBIM proxy could be constrained
to run as a less-privileged user and specially granted the permission to
access the MBIM device. So, adding a compile time flag in libmbim to check
for the specified user, rather than assume it to be the root user. If the flag is
not sent, it'll revert to the existing behaviour of checking for
user=root(i.e UID=0)