libqmi-glib,proxy: add a configure flag to set the user ID of QMI proxy
Currently, the QMI proxy process assumes that it is run as root user and
that all incoming client connection users are also root.
However, it's not always preferable to run the QMI proxy as root for
security reasons. On some platforms, the QMI proxy could be constrained
to run as a less-privileged user and specially granted the permission to
access the QMI device. So, adding a compile time flag in libqmi to check
for the specified user, rather than assume it to be the root user.  If the flag
is not sent, it'll revert to the existing behaviour of checking for
user=root(i.e UID=0)