libqmi-glib: always attach sources to the thread default main context
If NULL is specified in g_source_attach() it will attach to the default context,
not to the thread-default one if one was given with
g_main_context_push_thread_default().

This caused that QmiDevices started in the non-main thread would still attach
their socket listening sources to the GMainContext in the main thread.