qmi-codegen: fix public struct type generation
This change triggers an API break.

When building structs to be included in the public header, we were just relying
on using the 'public_format' of each variable. This is an error, as the variable
may be more complex than just public/private. E.g. could be another struct, or
an array, or a fixed sized string, as in the example.

In particular, this bug currently affects one public type, where one of its
elements changes from being just a pointer to a string to a fixed sized array of
4 bytes.

The following type is changed from:

    typedef struct _QmiMessageNasGetCellLocationInfoOutputGeranInfoCellElement {
        guint32 cell_id;
        gchar * plmn;
        guint16 lac;
        guint16 geran_absolute_rf_channel_number;
        guint8 base_station_identity_code;
        guint16 rx_level;
    } QmiMessageNasGetCellLocationInfoOutputGeranInfoCellElement;

To:

    typedef struct _QmiMessageNasGetCellLocationInfoOutputGeranInfoCellElement {
        guint32 cell_id;
        gchar plmn[4];
        guint16 lac;
        guint16 geran_absolute_rf_channel_number;
        guint8 base_station_identity_code;
        guint16 rx_level;
    } QmiMessageNasGetCellLocationInfoOutputGeranInfoCellElement;

Thanks to Joseba Sanjuan <joseba.sanjuan@gmail.com> for finding the bug.