libmbim-glib,message: new helper method to parse responses
The MBIM "Command", "Open" and "Close" request messages can be replied using
either a type specific response (e.g. "Command Done", or "Open Done" or "Close
Done" or instead using a generic "Function Error" message.

In order to ease the users of the library to look for errors in the received
responses, the new mbim_message_response_get_result() method expects
any of those two previous messages.

We also bump version of the library to 1.11.1, to indicate when the new API
was introduced, but this is Not a real release.