qmi-codegen: avoid breaking API when defining strings in public structs
Commit b9c3701e337198 introduced an API break, where we would change a pointer
to a heap allocated string and instead use a fixed size char array. This commit
will instead recover the pointer to the string when used in a public struct, so
that API isn't broken w.r.t. previous stable libqmi versions.

The string is now properly allocated before reading and deallocated as part of
the struct deallocation.