mbim-proxy: upon new client, always check if device really open with a command
Never assume that mbim_device_is_open() will return the valid open status of the
real device. If we're in the middle of a suspend/resume cycle, the logic will
think that the session is open, while in reality it may not be.

To really make sure that we're open, explicitly send a Command message to the
device, and react on the possible NotOpened error by re-opening the device right
away.