nas: fix ECIO List TLV array element type (api break)
ECIO List TLV (0x12) in the NAS Get Signal Strength response message is given as an
unsigned byte, representing negative 0.5dB increments.

This is an API break; added to NEWS already not to forget about it.

Also, version bump to 0.13.2 to flag the exact version in which the change
was introduced.

https://bugs.freedesktop.org/show_bug.cgi?id=85052