qmi-codegen: avoid variable names like 'new'
For C++ compatibility, we should not generate variables with the names of
protected C++ keywords, like 'new'. Currently this was happening in the "DMS Set
Service Programming Code" method, where we had a TLV named "New". Instead of
renaming the TLV, which would change the name of the getter method as well, we
will force prepending the "value_" string to the variable names of the getter
and setter methods.

https://bugs.freedesktop.org/show_bug.cgi?id=91479