mbim-codegen: fix message name in generated code doc for parse functions
At the point where the codegen happens, {name} is the name of the last
function argument, not of the message.  Also fix up the documentation to
say "Parses and returns parameters" instead of "Create a new request" since
parsing doesn't create any requests.

New: * Parses and returns parameters of the 'Signal State' notification command in the 'Basic Connect' service.
Old: * Create a new request for the 'ErrorRateThreshold' notification command in the 'Basic Connect' service.