libqmi-glib,device: fix segfault when cancellable already cancelled
    Thread 1 received signal SIGSEGV, Segmentation fault.
    0x00007ffff79c9105 in transaction_cancelled (cancellable=0x7fffe4009420, ctx=0x669a30) at qmi-device.c:268
    268	    tr->cancellable_id = 0;

The g_cancellable_connect() method will also call the given callback
when the input cancellable is already cancelled. This means that the
cancellation callback should also handle the case where the transaction
hasn't been stored in the tracking table yet.

Thanks to Benoît Donnette <benoit.donnette@21net.com> for the report and
the suggested fix.

https://bugs.freedesktop.org/show_bug.cgi?id=98283