qmi-firmware-update: new --override-download option
The Set Firmware Preference command may return an empty list of images
to be downloaded. If that happens and we just power cycle, we won't get
into QDL download mode, the module will just apply the new firmware
selection preference.

With the new --override-download option, we tell the module that even if
we already have the images, we want to perform the download.

This option doesn't apply to older SWI9200 devices as these don't have
any firmware preference setting. The option is implicit for these
devices, though, the download always happens.