libqmi-glib: support vendor-specific request/responses
We want to support non-standard messages that may be encoded with
different TLVs depending on how the vendor implemented them.

Anyway, right now this is really just to support the correct translation
of TLVs and message contents in the get_printable() methods.

The support is only included for QMI request/responses, and not for QMI
indications. This is because the library knows in which moment the
requests are created (and can apply the same rules to the matched
response when it is received). For the indications, though, there is no
such context configurable yet.