libqmi-glib,device: new qmi_device_close_async() operation
The sync close operation that was implemented was partially useless
because the source responsible for I/O was scheduled in the main
context, so we were really relying on that main context to be still
running to properly finish the operation. i.e. a qmi_device_close()
after the default main loop was stopped would always end up with a
timeout error, as the response was never read.

So, we now modify the sync operation so that it doesn't do any explicit
wait for the response, we just issue the MBIM close command.

And we provide a new async operation that may be used to wait for the
MBIM close response.