qmi-firmware-update: explicitly async-close QmiDevice before exiting
After an update operation, before exiting, we MUST close the QmiDevice
using the asynchronous method, so that we wait for the inner 'MBIM
close' message (when QMI over MBIM is being used).

We don't want to leave the 'MBIM close done' message unread, as that
seems to interfere with the next 'MBIM open' sequence.

This sequence is incorrect:

    qmi-firmware-update
      ...
      --> mbim close
      (exit without reading response)

    qmi-firmware-update
      --> mbim open (seq 1)
      <-- mbim close done
      --> mbim open (seq 2)
      --> mbim open (seq 3)
      --> mbim open (seq 4)
      ...
      (times out)

This sequence is correct:

    qmi-firmware-update
      ...
      --> mbim close
      <-- mbim close done
      (exit)

    qmi-firmware-update
      --> mbim open (seq 1)
      <-- mbim open done
      ...