libmbim-glib: add MBIM_READY_INFO_FLAG_NONE to MbimReadyInfoFlag
According to MBIM spec Rev 1.0 Errata-1 Table 10-17, a value 0 of the
ReadyInfo field in MBIM_SUBSCRIBER_READY_INFO refers to
MBIMReadyInfoFlagsNone and indicates that the device is in normal mode.

This patch adds MBIM_READY_INFO_FLAG_NONE (0) to the MbimReadyInfoFlag
enum, such that `mbimcli --query-subscriber-ready-info` shows 'Ready
info' as 'none' instead of 'unknown' when the ReadyInfo field is 0.