device: avoid signals sent to the qmi-proxy process
If e.g. qmi-proxy is started by ModemManager and we send a Ctrl+C to
it, the signal would be propagated to the qmi-proxy process and we
would kill it right away, while leaving ModemManager still around
wondering why the socket to the proxy got a HUP.

Avoid this, by making sure the qmi-proxy gets its own process group.