libqmi-glib,device: allow enabing QMI indications when in MBIM mode
The QMI indications via MBIM need to be explicitly enabled using the
"Set Subscribe List" MBIM command. We define a new open flag that
allows us telling the QMI stack that the generated device should
expect indications.

We don't do this unconditionally because a lot of times we're just
running QMI requests+responses with qmicli, and so the extra
configuration step wouldn't be required for anything.