qmicli: make PDC and LOC operations work via QMI over MBIM
Most PDC and LOC operations are based on receiving QMI indications
from the device. When using QMI over MBIM, the support for QMI
indications must be explicitly enabled, otherwise none will be
received.

We now explicitly tell the stack that we "expect indications", which
for the generic QMI-only case will do nothing but for the
QMI-over-MBIM case will do the explicit enabling required.