libqmi-glib,device: avoid closing same file descriptor twice
The GUnixInputStream and GUnixOutputStream object created in
create_iostream_with_fd() references the same file descriptor and are
both set to close the file descriptor when the stream is closed. That
leads to the same file descriptor being closed twice, which isn't
desirable. This patch addresses the issue by having the QmiDevice object
to keep track of the file descriptor and close it, which avoids
any potential race condition in the future.

The issue was observed by Eric Caruso <ejcaruso@chromium.org>