libqmi-glib,message: allow empty TLVs
String TLVs, i.e. TLVs that contain a single string and therefore
don't need any explicit size prefix, may be empty if they're
containing an empty string. This means that the TLV would have 0 bytes
as payload, and therefore we shouldn't treat that as an error.

The QMI_CORE_ERROR_TLV_EMPTY error type, introduced in 7a0f4856 for
libqmi 1.12, is reported as deprecated in the documentation, as it
will no longer be generated by the library.