mbim-codegen: remove broken support for guint32-array types out of struct
The only 2 fields flagged with 'guint32-array' type are given inside
structs, and this lead to having a buggy implementation of the type
when out of structs. E.g. we were assuming we could use '${public}' to
refer to the enum/flag type associated to the guint32 when in an
array, but that logic was totally broken as the public variable would
be equal to 'guint32-array' if no enum/flag type given. Just remove
that, and leave only the in-struct support.