libmbim-glib,proxy: full event entry cleanup on NotOpened function error
If the device suddenly reports that the MBIM session is not open,
e.g. after a suspend/resume cycle, we also cannot assume that the
client-configured lists of events are valid.

E.g. if the modem had QMI-over-MBIM indications enabled before the
suspend operation, once the modem is re-configured after the resume we
wouldn't be re-enabling these indications because we wrongly thought
they were already enabled.

This fixes the QMI-over-MBIM LOC support, which was broken after a
suspend/resume cycle.