libmbim-glib: explicitly configure only non-standard services
The MBIM documentation states that the DEVICE_SERVICE_SUBSCRIBE_LIST
message may be used by clients to specify which indications it is
interested in receiving, and then the modem "must only indicate
notifications for CIDs which have been enabled via this command".

A bit later it says that "Upon OPEN_DONE completion, notifications for
CIDs defined in this specification are enabled (aka subscribed to) by
default, and vendor extension notifications are off by default."

These two previous statements seem to contradict each other, as this
would mean that devices emit by default indications for all standard
services, regardless of whether clients have subscribed to them
explicitly via DEVICE_SERVICE_SUBSCRIBE_LIST.

Looking at actual real devices, giving a DEVICE_SERVICE_SUBSCRIBE_LIST
that includes NO basic service CID enabled does NOT disable all
indications of the standard services, which means we have no way to
do so.

This commit simplifies the "merged" service subscribe list handling by
making the assumption that basic services are by default already
subscribed to and also that there is no way to un-subscribe from
them. We will use DEVICE_SERVICE_SUBSCRIBE_LIST only to track the
subscriptions to CIDs of non-standard services in the device.

If the client subscribes to standard services explicitly, the proxy
will ignore those and will not send those explicit subscriptions to
the device. In the same way, if the client un-subscribes from standard
services it was previously subscribed to, we will also ignore those
attempts. Clients should be simplified and ignore these subscriptions,
but meanwhile the proxy should really behave as the real device.

And while trying to behave as the real device, the indications emitted
for standard services should always be sent to the clients, even if
the clients tried to do custom management of which CIDs of the
standard services it's subscribed to. We shouldn't add additional
filtering in our end if the device isn't doing that.