ms-basic-connect-extensions: rename service
This service is not a generic service, it is defined by Microsoft, so
rename it to follow the same naming scheme as for other services also
defined by Microsoft.

We do not provide compatibility symbols, as this is a NEW service to
be first released in libmbim 1.18.0.

We do bump the micro versionto 1.17.4, so that applications relying
on the git version of libmbim can bump their minimum required version
accordingly.