libmbim-glib,proxy: also track standard services
The assumption that we could totally ignore enabling/disabling
standard services seems to be wrong, as not all devices follow that
logic. Looks like some do expect the FULL list of service/cids to
subscribe to on every request, and we are not specifying the standard
services, we would totally be losing the notifications they emit.

So, track the full default list of CIDs of all standard services that
emit notifications, and assume that all new clients have that list
enabled by default. At device level we will NEVER disable the standard
services, but clients may do so. This means that the list of
service/cids that we pass down to the device must ALWAYS include all
cids from all standard services.

Fixes: 0073cbed59a270e1a8f18f591d31cbaf64a995ae