qmi-message: add functions to operate on QMI part of QmiMessage
QRTR-based transports will not have QMUX headers, and will instead
encode that information in a different way. However, that data will
still need to be passed around as part of the message by the library
consumer, so reusing the QMUX header structure to pass it around
should be fine and we will need to be able to create fake QMUX
headers.

Also add a unit test for the new functions.