build: Rename template files
The template files in the `build-aux/templates` directory used in
enums sources and headers generation use the `{$name}-template.[ch]`
pattern. However, the examples in the official guide[0] and usually
GNOME and Freedesktop packages use the `{$name}.[ch].template`
pattern.

Due to this, the name of the template files and build commands have
been changed.

[0] https://developer.gnome.org/gobject/stable/glib-mkenums.html