qmi-endpoint: begin moving transport hangup handler
Since the endpoint will eventually abstract away the transport,
it should be the one to alert the QmiDevice when the transport
hangs up. Since the transport is still part of the device for now
this is done with a bunch of callbacks but future patches will
get rid of __qmi_endpoint_hangup in favor of tracking that all
internally.

Also fixes an issue where a hangup by the MBIM device would go
unnoticed by the QmiDevice.