qmi-endpoint: take ref to QmiFile and use it for transport
This allows the QmiEndpoint to see what file it needs to open
streams for. Also prepare for opening different types of Endpoints
by moving the creation to device_open_step, after the driver has
been decided.

Because we have the file we can now add the object hierarchy for
the different implementations of the QmiEndpoint. Currently this
is regular QMUX or MBIM.