libqmi-glib,device: setup SYNC retries
If the application opens the QmiDevice just with the SYNC flag and not
with VERSION_CHECK, then there is no retry mechanism in place, and the
initial SYNC attempt sent to the modem would timeout and fail the open
operation.

Setup some minor retry mechanism applicable to the SYNC operation, so
that we're able to attempt running it several times in a row before
returning an open error.

This fixes the qmi-firmware-update operation with devices that ignore
the first QMI message sent to them, like the EM7565.