libqmi-glib,device: new qmi_device_command_abortable() method
This new method should be used when a command is abortable and we want
to wait for the abort operation to finish before returning the result
of the operation.

This logic solves race condition issues when e.g. a response to the
original request arrives just after we have sent the abort, so the
abort ends up failing. In this situation we should return the
successful response to the user and forget about the internal abort
sequence.