qmi-firmware-update: fix floating point exception with MC7455
If we attempt to run qmi-firmware-update on the MC7455, we first need
to detect that the device is NOT a sahara device, and in order to do
that we'll just try to create a QfuSaharaDevice.

During this process, we will try to read from the ttyUSB0 the 'hello'
message that the device sends to us, but it may happen that the device
does not send anything at all and we would end up reading 0 bytes. If
this happens, we should treat it as an indication that this is not a
sahara device right away. If we don't do this, the protocol runner
sahara_device_run_protocol_step() method will return STEP_UNKNOWN
without any error set, which ends up making sahara_device_initialize()
return FALSE without error, and that will make initable_init() wrongly
succeed.

  [14 Aug 2019, 20:28:08] [Debug] [qfu-sahara-device] opening TTY: /dev/ttyUSB0
  [14 Aug 2019, 20:28:08] [Debug] [qfu-sahara-device] setting terminal in raw mode...
  [14 Aug 2019, 20:28:08] [Debug] [qfu-sahara-device] waiting time for device to boot properly...
  [14 Aug 2019, 20:28:10] [Debug] [qfu-sahara-device] initializing sahara protocol...
  [14 Aug 2019, 20:28:13] [Debug] [qfu-updater] selected file 'SWI9X30C_02.26.01.00.cwe' (64480431 bytes)
  downloading cwe image: SWI9X30C_02.26.01.00.cwe (64.5 MB)...
  Floating point exception (core dumped)

  #0  0x0000000000410a8b in qfu_sahara_device_firehose_setup_download (self=0x2110680, image=0x2124dc0, n_blocks=0x7ffc3baf5838,cancellable=0x211c240, error=0x7ffc3baf5860) at qfu-sahara-device.c:573
  #1  0x0000000000407b89 in download_image_firehose (device=0x2110680,#image=0x2124dc0, cancellable=0x211c240, error=0x7ffc3baf5860) at qfu-updater.c:691
  #2  0x00000000004080d9 in run_context_step_download_image (task=0x2125840) at qfu-updater.c:808
  #3  0x0000000000408595 in run_context_step (task=0x2125840) at qfu-updater.c:1490
  #4  0x00000000004077c9 in run_context_step_cb (task=0x2125840) at qfu-updater.c:366
  #5  0x00007fd1e0f0becc in g_main_context_dispatch () from /usr/lib/libglib-2.0.so.0
  #6  0x00007fd1e0f0c238 in g_main_context_iterate.isra () from /usr/lib/libglib-2.0.so.0
  #7  0x00007fd1e0f0c492 in g_main_loop_run () from /usr/lib/libglib-2.0.so.0
  #8  0x00000000004067c8 in operation_update_run (updater=0x211c210, images=0x211b1f0) at qfu-operation-update.c:108
  #9  0x0000000000406862 in qfu_operation_update_download_run (images=0x211b1f0,#device_selection=0x211b820) at qfu-operation-update.c:163
  #10 0x0000000000405f3f in main (argc=1, argv=0x7ffc3baf5bf8) at qfu-main.c:726